extract the donateButtonRequest and make it a separate function to use in 2 components
