+++ src/app/common/constants/custom-types.constants.ts
@@ -1,0 +1,1 @@
+export type PositiveNumber = number & { _brand: 'PositiveNumber' };
