+++ src/features/AdditionalPages/SupportUsPage/components/DonationBlock.component.tsx
@@ -1,13 +1,12 @@
 import './DonationBlock.styles.scss';
 
 import { ChangeEvent, useEffect, useState } from 'react';
+import donateButtonRequest from '@app/common/requests/donateButtonRequest';
+import { PositiveNumber } from '@constants/custom-types.constants';
 
 import { Button, Checkbox } from 'antd';
 
-import DonationApi from '@/app/api/donates/donation.api';
 import useWindowSize from '@/app/common/hooks/stateful/useWindowSize.hook';
-import { donateEvent } from '@/app/common/utils/googleAnalytics.unility';
-import Donation from '@/models/feedback/donation.model';
 
 // eslint-disable-next-line complexity
 const DonationBlock = () => {
@@ -55,19 +54,8 @@ const DonationBlock = () => {
     const style = { '--input-width': `${inputWidth}px` } as React.CSSProperties;
 
     const handlePost = () => {
-        const donation: Donation = {
-            amount: donateAmount,
-            pageUrl: window.location.href,
-        };
-
-        if (isCheckboxChecked) {
-            const windowReference = window.open() as Window;
-            DonationApi.create(donation).then(({ pageUrl }) => {
-                donateEvent('support_us_page_donation_block');
-                windowReference.location = pageUrl;
-            }).catch((err) => {
-                windowReference.close();
-            });
+        if (isCheckboxChecked && donateAmount > 0) {
+            donateButtonRequest(donateAmount as PositiveNumber);
         }
     };
